requirements.txt
asgiref==3.7.2
Django==3.2.25
django-filter==23.5
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.7.1
